pyproject.toml
[project]
name = "video-image-resizing"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "pillow>=12.1.1",
]

[dependency-groups]
dev = [
    "pillow>=12.1.1",
    "pytest>=9.0.2",
]
